Command-ified the PlayerController
Refactored the PlayerController with the Command pattern to support the recording of previous play input for data analytics testing purposes
