+++ Assets/Code/Scripts/Player/Input/Commands/DeployCommand.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Code.Scripts.Player.Input.Commands
+{
+    public class DeployCommand : ICommand
+    {
+        private Pawn _controlledObject;
+
+        public DeployCommand(Pawn controlledObject)
+        {
+            _controlledObject = controlledObject;
+        }
+
+        public void Execute()
+        {
+            _controlledObject.DeployTrap();
+        }
+    }
+}
