+++ Assets/Code/Scripts/Player/Input/Commands/DeployCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa6fba272c686f54a9ccf67496ac5e90
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
