+++ Assets/Code/Scripts/Player/Input/Commands/IdleCommand.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Code.Scripts.Player.Input.Commands
+{
+    public class IdleCommand : ICommand
+    {
+        private Pawn _controlledObject;
+
+        public IdleCommand(Pawn controlledObject)
+        {
+            _controlledObject = controlledObject;
+        }
+
+        public void Execute()
+        {
+            _controlledObject.StandIdle();
+        }
+    }
+}
