+++ Assets/Code/Scripts/Player/Input/Commands/IdleCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 06a8137dc9359694597a32922f1e4df7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
