+++ Assets/Code/Scripts/Player/Input/Commands/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c83dc28678655b5419552bf0c7fe14c2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
