+++ Assets/Code/Scripts/Player/Input/Commands/Interfaces/ICommand.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Code.Scripts.Player.Input.Commands
+{
+    public interface ICommand
+    {
+        public void Execute();
+    }
+}
