+++ Assets/Code/Scripts/Player/Input/Commands/Interfaces/ICommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99ebe32448665aa488709c1d678ae913
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
