+++ Assets/Code/Scripts/Player/Input/Commands/JumpCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b328492e300eccf4ab62d1541df51bcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
