+++ Assets/Code/Scripts/Player/Input/Commands/MoveCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 798e45e25d0ee2245aa32af83bcbef3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
