+++ Assets/Code/Scripts/Player/Input/Commands/NullCommand.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Code.Scripts.Player.Input.Commands
+{
+    // Null Object Command
+    public class NullCommand : ICommand
+    {
+        public void Execute()
+        {
+            
+        }
+    }
+}
