+++ Assets/Code/Scripts/Player/Input/Commands/SetTrapCommand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1dd6e30599635374aac37770070d8ba5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
