+++ Assets/Code/Scripts/Player/Input/Pawn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9054cff65897174abe6ec43d8914066
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
