+++ Assets/Code/Scripts/Player/Input/PlayerStates/GameOverState.cs
@@ -1,7 +1,9 @@
 using System;
+using System.Collections.Generic;
+using Code.Scripts.Player.Input.Commands;
 using UnityEngine;
 
-namespace Input {
+namespace Code.Scripts.Player.Input {
     /// <summary>
     /// TODO: Make note of any music plugins we need here...
     /// </summary>
@@ -22,7 +24,7 @@ namespace Input {
             return 0f;
         }
         
-        public void Act(Rigidbody2D rBody, float direction, Action enterIdle)
+        public void Act(PlayerController playerController, float direction, List<ICommand> prevCommands)
         {
             // The game is over! Maybe the player shouldn't do anything anymore and some UI opens or something
         }
