+++ Assets/Code/Scripts/Player/Input/PlayerStates/MovingState.cs
@@ -1,7 +1,10 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using Code.Scripts.Player.Input.Commands;
 using UnityEngine;
 
-namespace Input {
+namespace Code.Scripts.Player.Input {
     /// <summary>
     /// TODO: Make note of any music plugins we need here...
     /// </summary>
@@ -26,9 +29,13 @@ namespace Input {
             return _stateSpeed;
         }
         
-        public void Act(Rigidbody2D rBody, float direction, Action enterIdle)
+        public void Act(PlayerController playerController, float direction, List<ICommand> prevCommands)
         {
-            rBody.velocity = new Vector2(direction * _stateSpeed, rBody.velocity.y);
+            // Create command and execute it
+            // Pass by reference in the list; need to create a new command each time to add to the list (Is there a
+            // better way of doing this, memory-wise?)
+            var command = new MoveCommand(playerController, direction);
+            playerController.ExecuteCommand(command);
         }
         
         public void OnExit(IPlayerState newState)
