--- Assets/Code/Scripts/Player/Input/PlayerStates/PlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 97d5f2f5a5eddb64cb93a1cfe96dc21d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
